+++ circuit_maintenance_parser/processor.py
@@ -116,17 +116,28 @@ class CombinedProcessor(GenericProcessor):
     combined_maintenance_data: Dict = {}
 
     def process_hook(self, maintenances_extracted_data, maintenances_data):
-        """All the parsers contribute with a subset of data that is extended."""
-        # We only expect one data object from these parsers
+        """All the parsers contribute with a subset of data that is extended.
+
+        For some notifications there can be multiple maintenances in a single file. To handle this, maintenances are store in a
+        list where each of them can be extended with the extra processors.
+        """
         if len(maintenances_extracted_data) == 1:
             self.combined_maintenance_data.update(maintenances_extracted_data[0])
         else:
-            raise ProcessorError(f"Unexpected data retrieved from parser: {maintenances_extracted_data}")
+            maintenances_data.extend(maintenances_extracted_data)
 
     def post_process_hook(self, maintenances_data):
         """After processing all the parsers, we try to combine all the data together."""
         self.extend_processor_data(self.combined_maintenance_data)
-        try:
-            maintenances_data.append(Maintenance(**self.combined_maintenance_data))
-        except ValidationError as exc:
-            raise ProcessorError("Not enough information available to create a Maintenance notification.") from exc
+        if not maintenances_data:
+            maintenances = [{}]
+        else:
+            maintenances = maintenances_data.copy()
+            maintenances_data.clear()
+
+        for maintenance in maintenances:
+            try:
+                combined_data = {**self.combined_maintenance_data, **maintenance}
+                maintenances_data.append(Maintenance(**combined_data))
+            except ValidationError as exc:
+                raise ProcessorError("Not enough information available to create a Maintenance notification.") from exc