+++ tests/unit/test_e2e.py
@@ -12,6 +12,7 @@ from circuit_maintenance_parser.errors import ProviderError
 # pylint: disable=duplicate-code
 from circuit_maintenance_parser.provider import (
     GenericProvider,
+    AquaComms,
     Cogent,
     Colt,
     EUNetworks,
@@ -22,6 +23,7 @@ from circuit_maintenance_parser.provider import (
     Momentum,
     PacketFabric,
     Seaborn,
+    Sparkle,
     Telia,
     Telstra,
     Turkcell,
@@ -40,6 +42,12 @@ GENERIC_ICAL_RESULT_PATH = Path(dir_path, "data", "ical", "ical1_result.json")
     [
         # GenericProvider
         (GenericProvider, [("ical", GENERIC_ICAL_DATA_PATH),], [GENERIC_ICAL_RESULT_PATH,],),
+        # AquaComms
+        (
+            AquaComms,
+            [("email", Path(dir_path, "data", "aquacomms", "aquacomms1.eml")),],
+            [Path(dir_path, "data", "aquacomms", "aquacomms1_result.json"),],
+        ),
         # Cogent
         (
             Cogent,
@@ -178,6 +186,12 @@ GENERIC_ICAL_RESULT_PATH = Path(dir_path, "data", "ical", "ical1_result.json")
             Seaborn,
             [("email", Path(dir_path, "data", "seaborn", "seaborn3.eml")),],
             [Path(dir_path, "data", "seaborn", "seaborn3_result.json"),],
+        ),
+        # Sparkle
+        (
+            Sparkle,
+            [("email", Path(dir_path, "data", "sparkle", "sparkle1.eml")),],
+            [Path(dir_path, "data", "sparkle", "sparkle1_result.json"),],
         ),
         # Telia
         (